Add an MVP Helm chart for using Rauthy in k8s.
This chart is meant to be a starting point. This chart will fully fire
up an instance of Rauthy in a k8s cluster. However, you will need some
form of smtp since rauthy crashes without it. This chart assumes the use
of mailcrab, which is a test smtp server that receives mail and present
a them in a web UI for testing.

The chart does not require a custom values yaml to launch and is
configured for testing. Any serious deployment would need to persist
some secrets that are generated in this chart.

Also, this chart runs rauthy on http inside the cluster. I am not sure
of a generic way to setup TLS for communication between the ingress
proxy and the services inside the cluster. If there is a good generic
solution that works be default with all ingress setups, I think that
would be a obvious next step.